refactor: improve ai chat organization
- Split AiChat into smaller components (MessageList, ChatInput)
- Created custom hook for chat logic (useAiChat)
- Extracted utilities and types
- Improved code organization with proper folder structure
- Added index file for exports
