+++ apps/web/src/components/ai-chat/ChatInput.tsx
@@ -1,0 +1,52 @@
+import { useEffect } from 'react';
+import { ChatInputProps } from './types';
+
+export function ChatInput({
+  input,
+  setInput,
+  isTyping,
+  onSubmit,
+  textareaRef
+}: ChatInputProps) {
+  // Auto-resize textarea as content grows
+  useEffect(() => {
+    const textarea = textareaRef.current;
+    if (textarea) {
+      textarea.style.height = 'auto';
+      textarea.style.height = Math.min(textarea.scrollHeight, 200) + 'px';
+    }
+  }, [input]);
+
+  const handleKeyDown = (e: React.KeyboardEvent<HTMLTextAreaElement>) => {
+    if (e.key === 'Enter' && !e.shiftKey) {
+      e.preventDefault();
+      onSubmit(e);
+    }
+  };
+
+  return (
+    <form onSubmit={onSubmit} className="border-t border-[#2d2d2d] p-4">
+      <div className="flex gap-2">
+        <textarea
+          ref={textareaRef}
+          value={input}
+          onChange={(e) => setInput(e.target.value)}
+          onKeyDown={handleKeyDown}
+          placeholder="Ask me anything..."
+          className="flex-1 bg-[#2d2d2d] rounded px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-500 resize-none min-h-[40px]"
+          disabled={isTyping}
+        />
+        <button
+          type="submit"
+          disabled={!input.trim() || isTyping}
+          className="bg-blue-600 text-white px-4 py-2 rounded text-sm hover:bg-blue-700 disabled:opacity-50 disabled:cursor-not-allowed transition-colors"
+        >
+          Send
+        </button>
+      </div>
+      <div className="mt-2 text-xs text-gray-400">
+        Press Shift + Enter for new line
+      </div>
+    </form>
+  );
+}
