+++ apps/web/src/components/ai-chat/index.ts
@@ -1,0 +1,6 @@
+export { AiChat } from './AiChat';
+export { MessageList } from './MessageList';
+export { ChatInput } from './ChatInput';
+export { useAiChat } from './useAiChat';
+export * from './types';
+export * from './utils';
