+++ apps/web/src/components/ai-chat/types.ts
@@ -1,0 +1,19 @@
+export interface Message {
+  role: 'assistant' | 'user';
+  content: string;
+  timestamp: Date;
+}
+
+export interface MessageListProps {
+  messages: Message[];
+  isTyping: boolean;
+  messagesEndRef: React.RefObject<HTMLDivElement>;
+}
+
+export interface ChatInputProps {
+  input: string;
+  setInput: (value: string) => void;
+  isTyping: boolean;
+  onSubmit: (e: React.FormEvent) => void;
+  textareaRef: React.RefObject<HTMLTextAreaElement>;
+}
