+++ apps/web/src/components/ai-chat/useAiChat.ts
@@ -1,0 +1,60 @@
+import { useState, useRef, useEffect } from 'react';
+import { Message } from './types';
+import { getInitialMessage, generateResponse } from './utils';
+
+export function useAiChat() {
+  const [messages, setMessages] = useState<Message[]>([getInitialMessage()]);
+  const [input, setInput] = useState('');
+  const [isTyping, setIsTyping] = useState(false);
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+  const textareaRef = useRef<HTMLTextAreaElement>(null);
+
+  const scrollToBottom = () => {
+    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+  };
+
+  useEffect(() => {
+    scrollToBottom();
+  }, [messages]);
+
+  const simulateAIResponse = (userMessage: string) => {
+    setIsTyping(true);
+    
+    setTimeout(() => {
+      const response = generateResponse(userMessage);
+      
+      setMessages(prev => [...prev, {
+        role: 'assistant',
+        content: response,
+        timestamp: new Date()
+      }]);
+      setIsTyping(false);
+    }, 1000);
+  };
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!input.trim() || isTyping) return;
+
+    const userMessage: Message = {
+      role: 'user',
+      content: input.trim(),
+      timestamp: new Date()
+    };
+
+    setMessages(prev => [...prev, userMessage]);
+    setInput('');
+    
+    simulateAIResponse(userMessage.content);
+  };
+
+  return {
+    messages,
+    input,
+    isTyping,
+    messagesEndRef,
+    textareaRef,
+    setInput,
+    handleSubmit
+  };
+}
